+++ cmd/stellar-rpc/internal/config/toml_test.go
@@ -113,7 +113,17 @@ func TestRoundTrip(t *testing.T) {
 		case *bool:
 			*v = true
 		case *string:
-			*v = "test"
+			switch option.ConfigKey {
+			case &cfg.Network:
+				// Network option sets the three config keys below, but requires them to be empty
+				cfg.HistoryArchiveURLs = []string{}
+				cfg.CaptiveCoreConfigPath = ""
+				cfg.NetworkPassphrase = ""
+				// Network option errors on inputs that are not a real network
+				*v = "testnet"
+			default:
+				*v = "test"
+			}
 		case *uint:
 			*v = 42
 		case *uint16: